RAG_apps/requirements.txt
langchain==0.1.10
langchain_community==0.0.25
langchain_core==0.1.29
langchain_openai==0.0.8
numpy==1.26.4
pandas==2.2.1
renumics-spotlight==1.6.7